+++ src/assessments/links/test-steps/link-function.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import * as React from 'react';
 
-import { LinkFunctionPropertyBag } from 'common/types/property-bag/ilink-function';
+import { LinkFunctionPropertyBag } from 'common/types/property-bag/link-function-property-bag';
 import { VisualizationType } from 'common/types/visualization-type';
 import { link } from 'content/link';
 import { title } from 'content/strings/application';
