+++ src/common/types/property-bag/link-function-property-bag.ts
